Search.php : max_num_pages doit être un entier.
Quand on fait max=ceil(total/size), on obtient un réel ce qui pose des soucis au code qui testent max_num_pages en utilisant l'égalité stricte (pager entre autres).

+++ class/Search.php
@@ -192,7 +192,7 @@ class Search extends Plugin {
         $size = $this->request->size();
 
         $query->found_posts = $total;
-        $query->max_num_pages = ceil($total / $size);
+        $query->max_num_pages = (int) ceil($total / $size);
 
         return $posts;
     }
